requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
pytest==8.3.2
pytest-dotenv==0.5.2
pytest-env==1.1.4
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.4
tomli==2.0.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
